move axios call to index
