--- client/src/QA.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import Question from './Question';
-
-function QA() {
-  return (
-    <div>
-      Q & A Module
-      <Question product_id={'40347'} />
-    </div>
-  );
-}
-
-export default QA;
